chore: Include Anonymization method in the Decomposition, remove unnecessary tests
